Move examples into top-level Sources directory, update API.
