--- Examples/Makefile
@@ -1,13 +1,0 @@
-
-
-all:
-	swift build
-	cp .build/debug/Twitter .
-	cp .build/debug/Spotify .
-	cp .build/debug/Google .
-	cp .build/debug/GitHub .
-	cp .build/debug/Meetup .
-
-clean:
-	rm -rf .build Twitter Spotify Google GitHub Meetup Package.pins *.json
-
