+++ Sources/Examples/Twitter/Twitter.swift
@@ -14,38 +14,21 @@
 
 import Foundation
 import Dispatch
-import OAuth2
+import OAuth1
 
-class SpotifySession {
+class TwitterSession {
+  public var connection : Connection
 
-  var connection : Connection
-
-  init(tokenProvider: TokenProvider) throws{
-    connection = try Connection(provider:tokenProvider)
+  init(tokenSource: TokenSource) throws{
+    connection = try Connection(source:tokenSource)
   }
 
-  func getUser() throws {
+  func getTweets() throws {
     let sem = DispatchSemaphore(value: 0)
     var responseData : Data?
-    connection.performRequest(
+    try connection.performRequest(
       method:"GET",
-      urlString:"https://api.spotify.com/v1/me") {(data, response, error) in
-        responseData = data
-        sem.signal()
-    }
-    _ = sem.wait(timeout: DispatchTime.distantFuture)
-    if let data = responseData {
-      let response = String(data: data, encoding: .utf8)!
-      print(response)
-    }
-  }
-
-  func getTracks() throws {
-    let sem = DispatchSemaphore(value: 0)
-    var responseData : Data?
-    connection.performRequest(
-      method:"GET",
-      urlString:"https://api.spotify.com/v1/me/tracks") {(data, response, error) in
+      urlString:"https://api.twitter.com/1.1/statuses/user_timeline.json") {(data, response, error) in
         responseData = data
         sem.signal()
     }
