+++ Sources/OAuth1/BrowserTokenSource.swift
@@ -22,13 +22,13 @@ struct AuthError : Error {
 
 }
 
-public class BrowserTokenProvider: TokenProvider {
+public class BrowserTokenSource : TokenSource {
+  public var consumerKey: String?
+  public var consumerSecret: String?
+  private var requestTokenURL: String?
   private var authorizeURL: String?
   private var accessTokenURL: String?
   private var callback: String?
-  private var clientID: String?
-  private var clientSecret: String?
-  private var code: Code?
   public var token: Token?
 
   public init(credentials: String, token tokenfile: String) throws {
@@ -43,16 +43,18 @@ public class BrowserTokenProvider: TokenProvider {
         case let .string(k):
           if let v = value.string {
             switch k {
+            case "consumer_key":
+              consumerKey = v
+            case "consumer_secret":
+              consumerSecret = v
+            case "request_token_url":
+              requestTokenURL = v
             case "authorize_url":
               authorizeURL = v
             case "access_token_url":
               accessTokenURL = v
             case "callback":
               callback = v
-            case "client_id":
-              clientID = v
-            case "client_secret":
-              clientSecret = v
             case "access_token_url":
               accessTokenURL = v
             default: break
@@ -74,7 +76,7 @@ public class BrowserTokenProvider: TokenProvider {
         }
         self.token = token
       } catch {
-        // ignore errors due to missing token files
+        // ignore errors due to missing session files
       }
     }
   }
@@ -91,13 +93,14 @@ public class BrowserTokenProvider: TokenProvider {
     // Create a new router
     let router = Router()
 
-    // Handle HTTP GET requests to the callback path
+    // Handle HTTP GET requests to /
     router.get(callback!) {
       request, response, next in
-      response.send("Hello!")
+      response.send("Hello, user!")
 
       let urlComponents = URLComponents(string: request.originalURL)!
-      self.code = Code(urlComponents: urlComponents)
+
+      self.token = Token(urlComponents: urlComponents)
       next()
 
       DispatchQueue.global().asyncAfter(deadline: .now() + 1) {
@@ -115,62 +118,64 @@ public class BrowserTokenProvider: TokenProvider {
     }
   }
 
-  private func exchange() throws -> Token {
+  public func signIn() throws {
+    let sem = DispatchSemaphore(value: 0)
+    startServer(sem: sem)
+
+    let sem2 = DispatchSemaphore(value: 0)
+
+    var parameters = ["oauth_callback": "http://localhost:8080" + callback!]
+    var responseData: Data?
+
+    Connection.performRequest(
+      method: "POST",
+      urlString: requestTokenURL!,
+      parameters: parameters,
+      tokenSecret: "",
+      consumerKey: consumerKey!,
+      consumerSecret: consumerSecret!) { data, _, _ in
+        responseData = data
+        sem2.signal()
+    }
+    _ = sem2.wait(timeout: DispatchTime.distantFuture)
+
+    let params = String(data: responseData!, encoding: .utf8)
+
+    var urlComponents = URLComponents(string: "http://example.com?" + params!)!
+    token = Token(urlComponents: urlComponents)
+
+    if true {
+      urlComponents = URLComponents(string: authorizeURL!)!
+      urlComponents.queryItems = [URLQueryItem(name: "oauth_token", value: encode(token!.oAuthToken!))]
+      openURL(urlComponents.url!)
+    }
+    _ = sem.wait(timeout: DispatchTime.distantFuture)
+    try exchange()
+  }
+
+  private func exchange() throws {
     let sem = DispatchSemaphore(value: 0)
     var parameters = [
-      "client_id": clientID!, // some providers require the client id and secret in the method call
-      "client_secret": clientSecret!,
-      "grant_type": "authorization_code",
-      "code": code!.code!,
-      "redirect_uri": "http://localhost:8080" + callback!,
+      "oauth_token": token!.oAuthToken!,
+      "oauth_verifier": token!.oAuthVerifier!,
       ]
-    let token = clientID! + ":" + clientSecret!
-    // some providers require the client id and secret in the authorization header
-    let authorization = "Basic " + String(data: token.data(using: .utf8)!.base64EncodedData(), encoding: .utf8)!
     var responseData: Data?
-    var contentType: String?
     Connection.performRequest(
       method: "POST",
       urlString: accessTokenURL!,
-      parameters: &parameters,
-      body: nil,
-      authorization: authorization) { data, response, _ in
-        if let c = (response as? HTTPURLResponse)!.allHeaderFields["Content-Type"] {
-          contentType = c as? String
-        }
+      parameters: parameters,
+      tokenSecret: "",
+      consumerKey: consumerKey!,
+      consumerSecret: consumerSecret!) { data, _, _ in
         responseData = data
         sem.signal()
     }
     _ = sem.wait(timeout: DispatchTime.distantFuture)
-    if contentType != nil && contentType!.contains("application/json") {
-
-      let decoder = JSONDecoder()
-      let token = try! decoder.decode(Token.self, from: responseData!)
-      return token
-    } else { // assume "application/x-www-form-urlencoded"
-      let urlComponents = URLComponents(string: "http://example.com?" + String(data: responseData!, encoding: .utf8)!)!
-      return Token(urlComponents: urlComponents)
-    }
+    let urlComponents = URLComponents(string: "http://example.com?" + String(data: responseData!, encoding: .utf8)!)!
+    token = Token(urlComponents: urlComponents)
   }
 
-  public func signIn(scopes: [String]) throws {
-    let sem = DispatchSemaphore(value: 0)
-    startServer(sem: sem)
-
-    let state = UUID().uuidString
-    let scope = scopes.joined(separator: " ")
-
-    var urlComponents = URLComponents(string: authorizeURL!)!
-    urlComponents.queryItems = [
-      URLQueryItem(name: "client_id", value: clientID!),
-      URLQueryItem(name: "response_type", value: "code"),
-      URLQueryItem(name: "redirect_uri", value: "http://localhost:8080" + callback!),
-      URLQueryItem(name: "state", value: state),
-      URLQueryItem(name: "scope", value: scope),
-      URLQueryItem(name: "show_dialog", value: "false"),
-    ]
-    openURL(urlComponents.url!)
-    _ = sem.wait(timeout: DispatchTime.distantFuture)
-    token = try exchange()
+  public func withToken(_ callback:@escaping (Token?, Error?) -> Void) throws {
+	callback(token, nil)
   }
 }
