+++ Sources/OAuth2/openURL.swift
@@ -19,7 +19,7 @@ import Foundation
 
 internal func openURL(_ url: URL) {
   #if os(OSX)
-    if !NSWorkspace.shared().open(url) {
+    if !NSWorkspace.shared.open(url) {
       print("default browser could not be opened")
     }
   #else
